Add bus and motorbike test

+++ src/Bus.java
@@ -1,6 +1,6 @@
 public class Bus extends Car {
     public Bus(String id) {
         this.id = ("bus_" + id);
-        length = (int) (super.getLength() * 3);
+        length = super.getLength() * 3;
     }
 }
