+++ src/Car.java
@@ -2,7 +2,7 @@ public class Car {
     private static final int STOPPED = 0; //car speed is 0m/s
     private static final int NEXT_ROAD_INDEX = 0;
     String id; // unique identifier
-    static int length; // number of segments occupied, 1 for ease of prototype.
+    static float length; // number of segments occupied, 1 for ease of prototype.
     private static float breadth;
     private int speed; //segments moved per turn
     private int[] position; // position on current road
